Implemented FileRegistry class to replace databases
Actually, the databases are more broken than you think. They rely on a
complex stack of many classes to work as expected and they also
incorporate too tight with the stored objects. For lglass 1.1, another
approach will be used. There will be a FileRegistry class for the old
file databases and a QueryEngine class to query a registry. Furthermore,
the algorithms will be tweaked and we will introduce a Digraph data
structure to project the foreign key relations in a database. By this
we will be able to provide a much better whois service and also a better
lglass-regtool. We aim to provide a usable regtool to replace manual
editing in registries.